+++ src/tests/healthz.test.ts
@@ -2,7 +2,10 @@ import { JSONResponse, Tile38 } from '..';
 
 describe('healthz', () => {
     describe('healthz leader', () => {
-        const tile38 = new Tile38(process.env.TILE38_URI);
+        const tile38 = new Tile38(
+            'redis://localhost:9851/',
+            'redis://localhost:9852/'
+        );
 
         afterAll(() => tile38.quit());
 
@@ -21,7 +24,10 @@ describe('healthz', () => {
     });
 
     describe('healthz follower', () => {
-        const tile38 = new Tile38(undefined, process.env.TILE38_URI);
+        const tile38 = new Tile38(
+            'redis://localhost:9851/',
+            'redis://localhost:9852/'
+        );
 
         afterAll(() => tile38.quit());